Made changes to program flow
Now some scripts will check to see if an instance of iteself is running. The prep-server script inserts the command to
the crontab if the backup env variable is included. Currently, the crontab is not executing as expect. Reasoning is
unknown and will need to be investigated further. All other scripts should be working as intended, now with the instance
checking enhancement. No breaking changes were observered.